refactor: fuse transformer implementations into unified StreamingMultiheadAttention
- Create streaming_attention.py with unified StreamingMultiheadAttention
- Add optional context parameter for sliding window attention
- Support both ring buffer (context window) and append (full sequence) KV caching
- Remove deprecated MimiStreamingMultiheadAttention class
- Convert transformer.py to backward compatibility shim
- All 12 tests pass with unified implementation

Closes #6
